refactor: vacation endpoint logic into infrastructure mediator
