+++ BookVacation.sln
@@ -100,8 +100,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Mediator", "Mediator", "{44
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VacationService.Infrastructure", "VacationService\VacationService.Infrastructure\VacationService.Infrastructure.csproj", "{BE2FBA6B-E34B-4A15-B539-6DE4E6320E3F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VacationService.Domain", "VacationService\VacationService.Domain\VacationService.Domain.csproj", "{89E74C53-DDE7-478B-87FF-37E921A84C3A}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VacationService.Tests", "VacationService\VacationService.Tests\VacationService.Tests.csproj", "{E08E6BC7-014F-4097-84CF-F06246C5FFA9}"
 EndProject
 Global
@@ -561,18 +559,6 @@ Global
 		{BE2FBA6B-E34B-4A15-B539-6DE4E6320E3F}.Release|x64.Build.0 = Release|Any CPU
 		{BE2FBA6B-E34B-4A15-B539-6DE4E6320E3F}.Release|x86.ActiveCfg = Release|Any CPU
 		{BE2FBA6B-E34B-4A15-B539-6DE4E6320E3F}.Release|x86.Build.0 = Release|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Debug|x64.Build.0 = Debug|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Debug|x86.Build.0 = Debug|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Release|x64.ActiveCfg = Release|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Release|x64.Build.0 = Release|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Release|x86.ActiveCfg = Release|Any CPU
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A}.Release|x86.Build.0 = Release|Any CPU
 		{E08E6BC7-014F-4097-84CF-F06246C5FFA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E08E6BC7-014F-4097-84CF-F06246C5FFA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E08E6BC7-014F-4097-84CF-F06246C5FFA9}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -623,7 +609,6 @@ Global
 		{982FF9DD-9190-481C-98DA-95B05055A49F} = {12B14152-34D1-4A0A-88F2-DFBDACC13B20}
 		{D16191CA-F809-4FB1-B3FF-7FBA1B344ECE} = {444FA222-26CB-48C4-AC4F-E739AE9A45DB}
 		{BE2FBA6B-E34B-4A15-B539-6DE4E6320E3F} = {17C4760F-7243-4FDB-AAF8-0FD757852377}
-		{89E74C53-DDE7-478B-87FF-37E921A84C3A} = {17C4760F-7243-4FDB-AAF8-0FD757852377}
 		{E08E6BC7-014F-4097-84CF-F06246C5FFA9} = {17C4760F-7243-4FDB-AAF8-0FD757852377}
 	EndGlobalSection
 	GlobalSection(RiderSharedRunConfigurations) = postSolution
