+++ VacationService/VacationService.Api/Dockerfile
@@ -9,6 +9,7 @@ WORKDIR /src
 COPY ["NuGet.config", "NuGet.config"]
 COPY ["VacationService/VacationService.Api/VacationService.Api.csproj", "VacationService.Api/"]
 COPY ["VacationService/VacationService.Contracts/VacationService.Contracts.csproj", "VacationService.Contracts/"]
+COPY ["VacationService/VacationService.Infrastructure/VacationService.Infrastructure.csproj", "VacationService.Infrastructure/"]
 RUN dotnet restore "VacationService.Api/VacationService.Api.csproj"
 COPY VacationService/ .
 WORKDIR "/src/VacationService.Api"
