+++ VacationService/VacationService.Api/VacationService.Api.csproj
@@ -9,23 +9,19 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="CarService.Contracts" Version="2022.3.24.2" />
-        <PackageReference Include="EventDispatcher" Version="2022.5.12.1" />
-        <PackageReference Include="FlightService.Contracts" Version="2022.3.24.1" />
         <PackageReference Include="FluentValidation.AspNetCore" Version="10.3.6" />
         <PackageReference Include="HealthCheck" Version="2022.5.28.1" />
-        <PackageReference Include="HotelService.Contracts" Version="2022.3.24.1" />
         <PackageReference Include="Logging" Version="2022.6.3.1" />
         <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning" Version="5.0.0" />
         <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning.ApiExplorer" Version="5.0.0" />
         <PackageReference Include="prometheus-net.AspNetCore" Version="5.0.2" />
         <PackageReference Include="prometheus-net.SystemMetrics" Version="2.0.0" />
         <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
-        <PackageReference Include="TicketService.Contracts" Version="2022.3.24.2" />
     </ItemGroup>
 
     <ItemGroup>
         <ProjectReference Include="..\VacationService.Contracts\VacationService.Contracts.csproj" />
+        <ProjectReference Include="..\VacationService.Infrastructure\VacationService.Infrastructure.csproj" />
     </ItemGroup>
 
 </Project>
