+++ VacationService/VacationService.Infrastructure/Requests/MediatorRegistration.cs
@@ -1,0 +1,6 @@
+namespace VacationService.Infrastructure.Requests;
+
+public struct MediatorRegistration
+{
+    
+}
