+++ VacationService/VacationService.Infrastructure/VacationService.Infrastructure.csproj
@@ -6,4 +6,15 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
+    <ItemGroup>
+      <PackageReference Include="CarService.Contracts" Version="2022.6.3.21" />
+      <PackageReference Include="EventDispatcher" Version="2022.6.3.3" />
+      <PackageReference Include="FlightService.Contracts" Version="2022.6.3.6" />
+      <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.0.2" />
+      <PackageReference Include="HotelService.Contracts" Version="2022.6.3.5" />
+      <PackageReference Include="Mediator" Version="2022.6.3.3" />
+      <PackageReference Include="MediatR" Version="10.0.1" />
+      <PackageReference Include="TicketService.Contracts" Version="2022.6.3.5" />
+    </ItemGroup>
+
 </Project>
