+++ VacationService/VacationService.Infrastructure/Validators/FluentValidatorRegistration.cs
@@ -1,0 +1,6 @@
+namespace VacationService.Infrastructure.Validators;
+
+public struct FluentValidatorRegistration
+{
+    
+}
